add hot swappable cluster algos

+++ tests/mockWeightTensors.py
@@ -10,7 +10,7 @@ class MockTensors:
                         [[[1, 2, 3], [4, 5, 6], [7, 8, 9]]],
                         [[[10, 11, 12], [13, 14, 15], [16, 17, 18]]],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "conv2.weight": torch.tensor(
                     [
                         [
@@ -22,7 +22,7 @@ class MockTensors:
                             [[28, 29, 30], [31, 32, 33], [34, 35, 36]],
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "fc1.weight": torch.tensor(
                     [
                         [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
@@ -45,7 +45,7 @@ class MockTensors:
                             32,
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
             },
             {
                 "conv1.weight": torch.tensor(
@@ -53,7 +53,7 @@ class MockTensors:
                         [[[9, 8, 7], [6, 5, 4], [3, 2, 1]]],
                         [[[18, 17, 16], [15, 14, 13], [12, 11, 10]]],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "conv2.weight": torch.tensor(
                     [
                         [
@@ -65,7 +65,7 @@ class MockTensors:
                             [[9, 8, 7], [6, 5, 4], [3, 2, 1]],
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "fc1.weight": torch.tensor(
                     [
                         [
@@ -88,7 +88,7 @@ class MockTensors:
                         ],
                         [16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1],
                     ]
-                ),
+                ,dtype=torch.float32),
             },
             {
                 "conv1.weight": torch.tensor(
@@ -96,7 +96,7 @@ class MockTensors:
                         [[[0, 1, 2], [3, 4, 5], [6, 7, 8]]],
                         [[[9, 10, 11], [12, 13, 14], [15, 16, 17]]],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "conv2.weight": torch.tensor(
                     [
                         [
@@ -108,7 +108,7 @@ class MockTensors:
                             [[45, 46, 47], [48, 49, 50], [51, 52, 53]],
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "fc1.weight": torch.tensor(
                     [
                         [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
@@ -131,7 +131,7 @@ class MockTensors:
                             32,
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
             },
             {
                 "conv1.weight": torch.tensor(
@@ -139,7 +139,7 @@ class MockTensors:
                         [[[2, 4, 6], [8, 10, 12], [14, 16, 18]]],
                         [[[20, 22, 24], [26, 28, 30], [32, 34, 36]]],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "conv2.weight": torch.tensor(
                     [
                         [
@@ -151,7 +151,7 @@ class MockTensors:
                             [[84, 87, 90], [93, 96, 99], [102, 105, 108]],
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "fc1.weight": torch.tensor(
                     [
                         [
@@ -191,7 +191,7 @@ class MockTensors:
                             320,
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
             },
             {
                 "conv1.weight": torch.tensor(
@@ -199,7 +199,7 @@ class MockTensors:
                         [[[1, 0, -1], [0, -1, -2], [-1, -2, -3]]],
                         [[[-4, -5, -6], [-7, -8, -9], [-10, -11, -12]]],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "conv2.weight": torch.tensor(
                     [
                         [
@@ -211,7 +211,7 @@ class MockTensors:
                             [[-28, -29, -30], [-31, -32, -33], [-34, -35, -36]],
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "fc1.weight": torch.tensor(
                     [
                         [
@@ -251,7 +251,7 @@ class MockTensors:
                             -32,
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
             },
             {
                 "conv1.weight": torch.tensor(
@@ -259,7 +259,7 @@ class MockTensors:
                         [[[5, 4, 3], [2, 1, 0], [-1, -2, -3]]],
                         [[[-4, -5, -6], [-7, -8, -9], [-10, -11, -12]]],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "conv2.weight": torch.tensor(
                     [
                         [
@@ -271,7 +271,7 @@ class MockTensors:
                             [[-17, -18, -19], [-20, -21, -22], [-23, -24, -25]],
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "fc1.weight": torch.tensor(
                     [
                         [15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0],
@@ -294,7 +294,7 @@ class MockTensors:
                             -16,
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
             },
             {
                 "conv1.weight": torch.tensor(
@@ -302,7 +302,7 @@ class MockTensors:
                         [[[1, 3, 5], [7, 9, 11], [13, 15, 17]]],
                         [[[19, 21, 23], [25, 27, 29], [31, 33, 35]]],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "conv2.weight": torch.tensor(
                     [
                         [
@@ -314,7 +314,7 @@ class MockTensors:
                             [[56, 58, 60], [62, 64, 66], [68, 70, 72]],
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "fc1.weight": torch.tensor(
                     [
                         [3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 33, 36, 39, 42, 45, 48],
@@ -337,7 +337,7 @@ class MockTensors:
                             96,
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
             },
             {
                 "conv1.weight": torch.tensor(
@@ -345,7 +345,7 @@ class MockTensors:
                         [[[2, 1, 0], [-1, -2, -3], [-4, -5, -6]]],
                         [[[-7, -8, -9], [-10, -11, -12], [-13, -14, -15]]],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "conv2.weight": torch.tensor(
                     [
                         [
@@ -357,7 +357,7 @@ class MockTensors:
                             [[-55, -57, -59], [-61, -63, -65], [-67, -69, -71]],
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "fc1.weight": torch.tensor(
                     [
                         [
@@ -397,7 +397,7 @@ class MockTensors:
                             -160,
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
             },
             {
                 "conv1.weight": torch.tensor(
@@ -405,7 +405,7 @@ class MockTensors:
                         [[[6, 5, 4], [3, 2, 1], [0, -1, -2]]],
                         [[[-3, -4, -5], [-6, -7, -8], [-9, -10, -11]]],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "conv2.weight": torch.tensor(
                     [
                         [
@@ -417,7 +417,7 @@ class MockTensors:
                             [[-15, -16, -17], [-18, -19, -20], [-21, -22, -23]],
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
                 "fc1.weight": torch.tensor(
                     [
                         [10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, -1, -2, -3, -4, -5],
@@ -440,6 +440,6 @@ class MockTensors:
                             -21,
                         ],
                     ]
-                ),
+                ,dtype=torch.float32),
             },
         ]
